+++ pkg/plugin/fields_query.go
@@ -15,6 +15,7 @@ type FieldsQuery struct {
 	Field              string `json:"field"`
 	ExtraFilters       string `json:"extra_filters"`
 	ExtraStreamFilters string `json:"extra_stream_filters"`
+	Filter             string `json:"filter"`
 }
 
 // getFieldsQueryFromRaw parses the field values query json from the raw message.
@@ -26,6 +27,7 @@ func getFieldsQueryFromRaw(data io.ReadCloser) (*FieldsQuery, error) {
 	return &q, nil
 }
 
+// API docs - https://docs.victoriametrics.com/victorialogs/querying/#querying-field-values
 func (fv *FieldsQuery) queryParams() url.Values {
 	params := url.Values{}
 	if fv.Query != "" {
@@ -48,6 +50,9 @@ func (fv *FieldsQuery) queryParams() url.Values {
 	}
 	if fv.ExtraStreamFilters != "" {
 		params.Set("extra_stream_filters", fv.ExtraStreamFilters)
+	}
+	if fv.Filter != "" {
+		params.Set("filter", fv.Filter)
 	}
 	return params
 }